refactor to ensure that what instantiating an indictor with fixed length source data no storage allocation occurs if there is not enough data to cover the lookback

+++ indicators/adl_test.go
@@ -64,7 +64,7 @@ var _ = Describe("when calculating an accumulation distribution line (adl) with 
 
 	Context("given the indicator is created via the constructor with fixed source length", func() {
 		BeforeEach(func() {
-			indicator, _ = indicators.NewAdlWithSrcLen(len(sourceDOHLCVData))
+			indicator, _ = indicators.NewAdlWithSrcLen(uint(len(sourceDOHLCVData)))
 			inputs = NewIndicatorWithFloatBoundsSharedSpecInputs(indicator, len(sourceDOHLCVData), indicator,
 				func() float64 {
 					return GetFloatDataMax(indicator.Data)
@@ -141,7 +141,7 @@ var _ = Describe("when calculating an accumulation distribution line (adl) with 
 	Context("given the indicator is created via the constructor for use with a price stream with fixed source length", func() {
 		BeforeEach(func() {
 			stream = newFakeDOHLCVStreamSubscriber()
-			indicator, _ = indicators.NewAdlForStreamWithSrcLen(len(sourceDOHLCVData), stream)
+			indicator, _ = indicators.NewAdlForStreamWithSrcLen(uint(len(sourceDOHLCVData)), stream)
 			inputs = NewIndicatorWithFloatBoundsSharedSpecInputs(indicator, len(sourceDOHLCVData), indicator,
 				func() float64 {
 					return GetFloatDataMax(indicator.Data)
