+++ indicators/adxr_test.go
@@ -154,7 +154,7 @@ var _ = Describe("when calculating an average directional movement rating (Adxr)
 
 	Context("given the indicator is created via the constructor with fixed source length", func() {
 		BeforeEach(func() {
-			indicator, _ = indicators.NewAdxrWithSrcLen(len(sourceDOHLCVData), 4)
+			indicator, _ = indicators.NewAdxrWithSrcLen(uint(len(sourceDOHLCVData)), 4)
 			inputs = NewIndicatorWithFloatBoundsSharedSpecInputs(indicator, len(sourceDOHLCVData), indicator,
 				func() float64 {
 					return GetFloatDataMax(indicator.Data)
@@ -193,7 +193,7 @@ var _ = Describe("when calculating an average directional movement rating (Adxr)
 
 	Context("given the indicator is created via the constructor with defaulted parameters and fixed source length", func() {
 		BeforeEach(func() {
-			indicator, _ = indicators.NewDefaultAdxrWithSrcLen(len(sourceDOHLCVData))
+			indicator, _ = indicators.NewDefaultAdxrWithSrcLen(uint(len(sourceDOHLCVData)))
 			inputs = NewIndicatorWithFloatBoundsSharedSpecInputs(indicator, len(sourceDOHLCVData), indicator,
 				func() float64 {
 					return GetFloatDataMax(indicator.Data)
@@ -307,7 +307,7 @@ var _ = Describe("when calculating an average directional movement rating (Adxr)
 	Context("given the indicator is created via the constructor for use with a price stream with fixed source length", func() {
 		BeforeEach(func() {
 			stream = newFakeDOHLCVStreamSubscriber()
-			indicator, _ = indicators.NewAdxrForStreamWithSrcLen(len(sourceDOHLCVData), stream, 4)
+			indicator, _ = indicators.NewAdxrForStreamWithSrcLen(uint(len(sourceDOHLCVData)), stream, 4)
 			inputs = NewIndicatorWithFloatBoundsSharedSpecInputs(indicator, len(sourceDOHLCVData), indicator,
 				func() float64 {
 					return GetFloatDataMax(indicator.Data)
@@ -351,7 +351,7 @@ var _ = Describe("when calculating an average directional movement rating (Adxr)
 	Context("given the indicator is created via the constructor for use with a price stream with fixed source length with defaulted parmeters", func() {
 		BeforeEach(func() {
 			stream = newFakeDOHLCVStreamSubscriber()
-			indicator, _ = indicators.NewDefaultAdxrForStreamWithSrcLen(len(sourceDOHLCVData), stream)
+			indicator, _ = indicators.NewDefaultAdxrForStreamWithSrcLen(uint(len(sourceDOHLCVData)), stream)
 			inputs = NewIndicatorWithFloatBoundsSharedSpecInputs(indicator, len(sourceDOHLCVData), indicator,
 				func() float64 {
 					return GetFloatDataMax(indicator.Data)
